Changed type of user input coordinates from uint to double

+++ src/cdpr_controller/src/UserNode.cpp
@@ -11,7 +11,7 @@ using std::cout;
 using std::cin;
 using std::endl;
 
-unsigned int x, y, xPrev, yPrev;
+double x, y, xPrev, yPrev;
 
 int getData() {
   cin >> x >> y;
